+++ xbmc/cores/paplayer/PAPlayer_xbox.cpp
@@ -782,9 +782,9 @@ __int64 PAPlayer::GetTotalTime64()
   return total;
 }
 
-int PAPlayer::GetTotalTime()
+int64_t PAPlayer::GetTotalTime()
 {
-  return (int)(GetTotalTime64()/1000);
+  return (int64_t)(GetTotalTime64()/1000);
 }
 
 int PAPlayer::GetCacheLevel() const
